Enhanced MediaType support and use Query param for providers
