+++ pom.xml
@@ -91,15 +91,6 @@
       <artifactId>jgrapht-io</artifactId>
       <version>${jgraph.version}</version>
     </dependency>
-    <dependency>
-      <groupId>org.projectlombok</groupId>
-      <artifactId>lombok</artifactId>
-      <version>${lombok.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.camel.quarkus</groupId>
-      <artifactId>camel-quarkus-log</artifactId>
-    </dependency>
 
     <!-- Test -->
     <dependency>
