+++ src/main/java/com/redhat/ecosystemappeng/crda/integration/VulnerabilityProvider.java
@@ -20,11 +20,14 @@ package com.redhat.ecosystemappeng.crda.integration;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.ExchangeProperty;
+import org.apache.camel.Header;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
@@ -35,7 +38,7 @@ public class VulnerabilityProvider {
 
     @ConfigProperty(name = "api.snyk.disabled", defaultValue = "false")
     private boolean snykDisabled;
-    
+
     @ConfigProperty(name = "api.tidelift.disabled", defaultValue = "false")
     private boolean tideliftDisabled;
 
@@ -44,21 +47,43 @@ public class VulnerabilityProvider {
     @PostConstruct
     public void initProviders() {
         providers = Collections.unmodifiableList(Constants.PROVIDERS.stream()
-            .filter(p -> !(Constants.SNYK_PROVIDER.equals(p) && snykDisabled))
-            .filter(p -> !(Constants.TIDELIFT_PROVIDER.equals(p) && tideliftDisabled))
-            .toList());
+                .filter(p -> !(Constants.SNYK_PROVIDER.equals(p) && snykDisabled))
+                .filter(p -> !(Constants.TIDELIFT_PROVIDER.equals(p) && tideliftDisabled))
+                .toList());
     }
 
     public String get(@ExchangeProperty(Exchange.MULTICAST_INDEX) int index) {
         return providers.get(index);
     }
 
-    public String[] getEndpoints() {
-        return providers.stream().map(p -> "direct:depAnalysis").toList().toArray(new String[] {});
+    public List<String> getProviderEndpoints(@ExchangeProperty(Constants.PROVIDERS_PARAM) List<String> providers) {
+        return providers.stream().map(p -> {
+            switch (p) {
+                case Constants.SNYK_PROVIDER:
+                    return "direct:snykDepGraph";
+                case Constants.TIDELIFT_PROVIDER:
+                    return "direct:tideliftReleases";
+                default:
+                    throw new IllegalArgumentException("Unexpected provider: " + p);
+            }
+        }).collect(Collectors.toList());
     }
 
     public List<String> getEnabled() {
         return providers;
+    }
+
+    public List<String> getProvidersFromQueryParam(@Header(Exchange.HTTP_QUERY) String query) {
+        if (query == null || query.trim().isBlank()) {
+            return getEnabled();
+        }
+        String[] params = query.split("&");
+        for (int i = 0; i < params.length; i++) {
+            if (params[i].startsWith(Constants.PROVIDERS_PARAM)) {
+                return List.of(params[i].replace(Constants.PROVIDERS_PARAM + "=", "").split(","));
+            }
+        }
+        return getEnabled();
     }
 
 }
