+++ src/main/java/com/redhat/ecosystemappeng/crda/integration/backend/BackendUtils.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * 
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.redhat.ecosystemappeng.crda.integration.backend;
+
+import java.util.List;
+import javax.ws.rs.ClientErrorException;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response.Status;
+
+import org.apache.camel.Header;
+import org.jboss.resteasy.reactive.common.util.MediaTypeHelper;
+
+import com.redhat.ecosystemappeng.crda.integration.Constants;
+
+import io.quarkus.runtime.annotations.RegisterForReflection;
+
+@RegisterForReflection
+public class BackendUtils {
+
+    public String getResponseMediaType(@Header("Accept") String acceptHeader) {
+        if (acceptHeader == null || acceptHeader.isBlank()) {
+            return Constants.DEFAULT_ACCEPT_MEDIA_TYPE;
+        }
+        List<MediaType> requested = MediaTypeHelper.parseHeader(acceptHeader);
+        MediaType match = MediaTypeHelper.getBestMatch(Constants.VALID_RESPONSE_MEDIA_TYPES, requested);
+        if (match == null) {
+            throw new ClientErrorException(
+                    "Unexpected Accept header " + acceptHeader + ". Supported content types are: "
+                            + Constants.VALID_RESPONSE_MEDIA_TYPES,
+                    Status.UNSUPPORTED_MEDIA_TYPE);
+        }
+        return match.toString();
+
+    }
+
+}
