+++ src/main/resources/META-INF/openapi.yaml
@@ -12,10 +12,30 @@ servers:
 - url: http://localhost:8080/api/v3/
   description: Local development
 paths:
-  /component-analysis:
+  /component-analysis/{pkgManager}:
     post:
       operationId: componentAnalysis
       summary: Performs a vulnerability for the provided components
+      parameters:
+      - name: pkgManager
+        in: path
+        description: Package manager related to the provided packages
+        required: true
+        schema:
+          enum:
+          - maven
+          type: string
+        example: maven        
+      - name: providers
+        in: query
+        description: List of vulnerability providers to get the report from
+        required: false
+        schema:
+          enum:
+          - snyk
+          - tidelift
+          type: string
+        example: snyk
       requestBody:
         content:
           application/json:
@@ -36,10 +56,30 @@ paths:
               schema:
                 type: string
                 description: Error message
-  /dependency-analysis:
+  /dependency-analysis/{pkgManager}:
     post:
-      operationId: fullDependencyAnalysis
+      operationId: dependencyAnalysis
       summary:  Takes a client-resolved dependency graph to perform a full stack analysis from all the available Vulnerability sources
+      parameters:
+      - name: pkgManager
+        in: path
+        description: Package manager related to the provided packages
+        required: true
+        schema:
+          enum:
+          - maven
+          type: string
+        example: maven
+      - name: providers
+        in: query
+        description: List of vulnerability providers to get the report from
+        required: false
+        schema:
+          enum:
+          - snyk
+          - tidelift
+          type: string
+        example: snyk
       requestBody:
         required: true
         description: Dependency graph in DOT format
@@ -64,61 +104,12 @@ paths:
             text/html:
               schema:
                 type: object
-  /dependency-analysis/{pkgManager}/{provider}:
-    post:
-      operationId: dependencyAnalysis
-      parameters:
-      - name: pkgManager
-        in: path
-        description: Package manager related to the provided packages
-        required: true
-        schema:
-          enum:
-          - maven
-          type: string
-        example: maven
-      - name: provider
-        in: path
-        description: Vulnerability provider to get the report from
-        required: true
-        schema:
-          enum:
-          - snyk
-          - redhat
-          type: string
-        example: snyk
-      responses:
-        '200': 
-          description: Successful component analysis report
-          content:
-            application/json:
-              schema: 
-                type: object
-        '422':
-          description: Invalid request
-          content:
-            text/plain:
-              schema:
-                type: string
-                description: Error message
 components:
   schemas:
     ComponentRequest:
-      type: object
-      properties:
-        pkgManager:
-          type: string
-          enum:
-            - maven
-        provider:
-          type: string
-          enum:
-            - snyk
-            - redhat
-        packages:
-          type: array
-          items:
-            $ref: '#/components/schemas/PackageRef'
+      type: array
+      items:
+        $ref: '#/components/schemas/PackageRef'
     PackageRef:
       type: object
       properties:
