+++ src/main/resources/application.properties
@@ -1,9 +1,8 @@
 %dev.quarkus.log.level=DEBUG
 
 ## podman run -p 18080:8080 --rm ghcr.io/seedwing-io/swio:0.1.0-trusted
-api.trustedContent.host=http://swio.trusted-content.svc.cluster.local:8080
-api.trustedContent.disabled=true
-api.trustedContent.vex.host=http://tc-camel.svc.cluster.local:8080
+api.trustedContent.gav.host=http://swio.trusted-content:8080
+api.trustedContent.vex.host=http://tc-camel:8080
 api.snyk.host=https://app.snyk.io
 # api.snyk.disabled=true
 api.tidelift.token=placeholder
